web/govpass/src/web/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
cryptography==42.0.8
requests==2.32.3
werkzeug==3.0.3
gunicorn==22.0.0
